% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abapToOccuR.R
\name{abapToOccuR}
\alias{abapToOccuR}
\title{ABAP to occuR}
\usage{
abapToOccuR(abap_data, occasion, pentads = NULL, proj_coords = TRUE)
}
\arguments{
\item{abap_data}{ABAP data downloaded using \code{\link{getAbapData}}.}

\item{occasion}{A character string indicating the variable in \code{abap_data}
that informs about the season (or occasion, as referred to in occuR). This must always
be supplied, although it is really only important in the case of multi-season data.
It is typically "year" but could be something else.}

\item{pentads}{Optional, An \code{sf} object returned by \code{\link{getRegionPentads}}. Defaults to \code{NULL}.}

\item{proj_coords}{logical value indicating whether pentad coordinates are
projected (\code{TRUE}) or kept in decimal degree format (\code{FALSE}). Defaults to \code{TRUE}.
See details below for coordinate reference system used during transformation.}
}
\value{
A list containing data necessary for model fitting in \code{occuR}.
List elements are: visit_data, a data frame containing information about individual
visits; site_data, a data frame containing information about sites and seasons.
If \code{pentads} are given then the coordinates of the centroid of the pentads will be
included in site_data.
}
\description{
This function transforms a raw ABAP data frame (returned by \code{\link{getAbapData}})
into an list which can be used to fit single-species occupancy models using
the package \code{\href{https://github.com/r-glennie/occuR}{occuR}}. This package
can fit non-linear effects, including spatial, temporal and spatio-temporal effects
using splines.
}
\details{
The \code{\href{https://github.com/r-glennie/occuR}{occuR}} package can
fit spatial effects, for which we need the spatial location of our sites. Within the context
of ABAP, these locations are the centroid of each sampling pentad. In order to
provide these spatial data to this function, simply use \code{\link{getRegionPentads}}
and provide the same inputs for \code{.region_type} and \code{.region} that are specified
in corresponding \code{\link{getAbapData}} call. If proj_coords is set to \code{TRUE}
then the coordinates will be transformed using the African Albers Equal Area
coordinate system (see \href{https://spatialreference.org/ref/esri/africa-albers-equal-area-conic/}{here}
for details). This projection is best suited for land masses extending in an
east-to-west orientation at mid-latitudes making it suitable for projecting
pentads in Southern Africa. \cr

In addition to reformatting the detection/non-detection ABAP data for use in
\code{occuR} occupancy models, this function also extracts two survey-level
covariates and adds them to the output list: \code{ hours} and \code{jday}. The \code{hours}
variable is the total number of hours spent atlassing which is recorded on the
pentad card and \code{jday} is the Julian day corresponding to the first day of
atlassing for that card.
}
\examples{
abap_data <- getAbapData(.spp_code = 212,
                         .region_type = "province",
                         .region = "Eastern Cape",
                         .years = 2012)

abap_pentads <- getRegionPentads(.region_type = "province",
                                 .region = "Eastern Cape")

# We will use years as occupancy reference seasons
abap_data$year <- format(abap_data$StartDate, "\%Y")

## Return list for spatial occupancy model
occur_data <- abapToOccuR(abap_data, occasion = "year", abap_pentads)
str(occur_data)

## Return list for spatial occupancy model (without coordinate projection)
occur_data <- abapToOccuR(abap_data, "year", abap_pentads, proj_coords = FALSE)
str(occur_data)

## List for non-spatial occupancy model
occur_data <- abapToOccuR(abap_data, "year")
str(occur_data)

}
