% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abapToSpOcc_single.R
\name{abapToSpOcc_single}
\alias{abapToSpOcc_single}
\title{ABAP to spOccupancy}
\usage{
abapToSpOcc_single(abap_data, pentads = NULL)
}
\arguments{
\item{abap_data}{ABAP data downloaded using \code{\link{getAbapData}}}

\item{pentads}{sf object returned by \code{\link{getRegionPentads}}. Defaults to \code{NULL}.}
}
\value{
A list containing data necessary for model fitting in \code{spOccupancy}. List elements are \code{y},  \code{det.covs}, and \code{coords} (if pentad data supplied)
}
\description{
Transforms a raw ABAP data frame into a list formated for use in an spOccupancy model
}
\note{
If pentads are specified then output can be used in \code{spOccupancy::spPGOcc}. If no spatial data are provided the output can be used in \code{spOccupancy::spPGOcc}
}
\examples{
abap_single <- getAbapData(.spp_code = 212, .region_type = "province",
                           .region = "Eastern Cape", .years = 2012)

abap_pentads <- getRegionPentads(.region_type = "province",
                                 .region = "Eastern Cape")
## List for spatial model
abapToSpOcc_single(abap_single, abap_pentads)

## List for non-spatial model
abapToSpOcc_single(abap_single)
}
