% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abapToUnmarked_single.R
\name{abapToUnmarked_single}
\alias{abapToUnmarked_single}
\title{ABAP to unmarked (single season)}
\usage{
abapToUnmarked_single(abap_data)
}
\arguments{
\item{abap_data}{ABAP data downloaded using \code{\link{getAbapData}}}
}
\value{
An object in the same format as that returned by \code{\link[unmarked]{unmarkedFrameOccu}}
}
\description{
Transforms a raw ABAP data frame into an unmarked-ready data frame
}
\note{
Total hours and Julian day are automatically added as observation-level covariates
}
\examples{
abap_single <- getAbapData(.spp_code = 212,
                          .region_type = "province",
                          .region = "Eastern Cape",
                          .years = 2012)

um_df <- abapToUnmarked_single(abap_single)
}
