% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVarEEcollection.R
\name{addVarEEcollection}
\alias{addVarEEcollection}
\title{Add variable from Google Earth Engine Collection}
\usage{
addVarEEcollection(
  ee_pentads,
  collection,
  dates,
  temp_reducer,
  spt_reducer,
  bands = NULL,
  unmask = FALSE
)
}
\arguments{
\item{ee_pentads}{A feature collection with the pentads we want to annotate.
We need to upload an sf object with the pentads to GEE.}

\item{collection}{Either a character string with the name of the collection
we want to use or a GEE collection produced with \code{ee$ImageCollection()}.
See \href{https://developers.google.com/earth-engine/datasets/catalog}{GEE catalog}.}

\item{dates}{A character vector with two elements c(start, end). Format must be
"yyyy-mm-dd".}

\item{temp_reducer}{A character string specifying the function to summarize
collection across time. It is common to use "mean", "sum" or "count", but
there are many others, see 'ee.Reducer' under Client Libraries at
\url{https://developers.google.com/earth-engine/apidocs}.}

\item{spt_reducer}{A character string specifying the function apply when
extracting values for each pentad. It is common to use "mean", "sum" or
"count", but there are many others, see 'ee.Reducer' under Client Libraries at
\url{https://developers.google.com/earth-engine/apidocs}.}

\item{bands}{Select specific bands from the image. If not specified, pentads
will be annotated with all bands (one column for each band).}

\item{unmask}{GEE masks missing values, which means they are not used for
computing means, counts, etc. Sometimes we might want to avoid this behaviour
and use 0 instead of NA. If so, set unmask to TRUE.}
}
\value{

}
\description{
Add variable from Google Earth Engine Collection
}
