% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVarFromEEclosestImage.R
\name{addVarFromEEclosestImage}
\alias{addVarFromEEclosestImage}
\title{Add variable from closest image ina Google Earth Engine Collection}
\usage{
addVarFromEEclosestImage(
  ee_pentads,
  collection,
  reducer,
  maxdiff,
  bands = NULL,
  unmask = FALSE
)
}
\arguments{
\item{ee_pentads}{A feature collection with the pentads we want to annotate.
We need to upload an sf object with the pentads to GEE. This object must
also have character column with the dates that need to be matched against the
image collection dates. The format must be "yyyy-mm-dd" and the column must
be named "Date".
#' @param reducer A character string specifying the function apply when
extracting values for each pentad. It is common to use "mean", "sum" or
"count". But there are many other, see 'ee.Reducer' under Client Libraries at
\url{https://developers.google.com/earth-engine/apidocs}.}

\item{collection}{The name of the collection we want to use. See
\href{https://developers.google.com/earth-engine/datasets/catalog}{GEE catalog}.}

\item{bands}{Select specific bands from the image. Only one band at a time is
allowed for now.}

\item{unmask}{GEE masks missing values, which means they are not used for
computing means, counts, etc. Sometimes we might want to avoid this behaviour
and use 0 instead of NA. If so, set unmask to TRUE.}
}
\value{

}
\description{
Each feature in a feature collection (i.e. row in asimple feature
collection) is matched with the image in an image collection that is closest
to it in time.
}
