% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPentadFromCoords.R
\name{findPentadFromCoords}
\alias{findPentadFromCoords}
\title{Find the ABAP pentad that contains a spatial location}
\usage{
findPentadFromCoords(x, y)
}
\arguments{
\item{x}{Longitude of the point contained by the pentad, in decimal degrees.}

\item{y}{Longitude of the point contained by the pentad, in decimal degrees.}
}
\value{
The code of the pentad containing the location defined by \code{x} and \code{y}.
}
\description{
Find the ABAP pentad that contains a spatial location
}
\examples{
findPentadFromCoords(-25.68606, 29.011337)
}
