% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCardRecords.R
\name{getCardRecords}
\alias{getCardRecords}
\title{Download the list of records from an ABAP card}
\usage{
getCardRecords(.CardNo)
}
\arguments{
\item{.CardNo}{An ABAP card code}
}
\value{
A tibble in which each row corresponds to an entry to an ABAP card.
}
\description{
Download the list of records from an ABAP card
}
\examples{
getCardRecords("3335_2320_011617_20211219")
}
