% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionPentads.R
\name{getRegionPentads}
\alias{getRegionPentads}
\title{Get ABAP pentads for a certain region}
\usage{
getRegionPentads(.region_type, .region)
}
\arguments{
\item{.region_type}{The type of region we are interested on.
Three options: "country", "province", "project" and "pentad".}

\item{.region}{A character string corresponding to the specific region we are
interested in. It can be either a country in Southern Africa, a South African
province, a project (currently, "SABAP1" or "SABAP2") or a pentad code.}
}
\value{
A simple feature object with those pentads that are contained in the
region of interest. A pentads is considered contained if half of its area
lies within the region.
}
\description{
Get ABAP pentads for a certain region
}
\details{
At the moment, ABAP serves data from eleven countries: South Africa,
Botswana, eSwatini, Lesotho, Malawi, Mozambique, Namibia, Zambia, Zimbabwe,
Kenya and Nigeria. Only South African provinces are available: Western Cape,
Eastern Cape, Northern Cape, Free State, Limpopo, Mpumalanga, KwaZulu-Natal,
Gauteng, and Northwest Province.
}
\examples{
# getRegionPentads(.region_type = "country", .region = "South Africa")
}
