% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonToTibble.R
\name{jsonToTibble}
\alias{jsonToTibble}
\title{Helper function to transform a JSON object to a tibble}
\usage{
jsonToTibble(jsonfile)
}
\arguments{
\item{jsonfile}{An object obtained when a JSON file is downloaded}
}
\value{
A tibble
}
\description{
Helper function to transform a JSON object to a tibble
}
\examples{
url <- "http://api.adu.org.za/cwac/sites/list?province=north20\%west"
myfile <- RCurl::getURL(url, ssl.verifyhost = FALSE, ssl.verifypeer = FALSE)
jsonfile <- rjson::fromJSON(myfile)
jsonToTibble(jsonfile)
}
