% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pentads_sabap.R
\docType{data}
\name{pentads_sabap}
\alias{pentads_sabap}
\title{Pentads of the SABAP2 project}
\format{
A simple feature with 60178 polygons and 1 variables:
\describe{
\item{Name}{Pentad code}
...
}
}
\usage{
pentads_sabap
}
\description{
A dataset containing all pentads considered in the SABAP2 project as simple features (sf)
}
\keyword{datasets}
