% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSabapSpecies.R
\name{searchSabapSpecies}
\alias{searchSabapSpecies}
\title{Search species in SABAP2 data base}
\usage{
searchSabapSpecies(species)
}
\arguments{
\item{species}{A character string with any part of the common or taxonomic name, or, the SABAP2 species code. Note that full names will not work; choose a single word.}
}
\value{
A tibble with taxonomic information and SABAP2 codes of all species matching the search
}
\description{
Search species in SABAP2 data base
}
\examples{
searchSabapSpecies("Duck")
searchSabapSpecies(169)
searchSabapSpecies("Fulvous")
}
