% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abapToSpOcc_single.R
\name{abapToSpOcc_single}
\alias{abapToSpOcc_single}
\title{ABAP to spOccupancy (single-species)}
\usage{
abapToSpOcc_single(abap_data, pentads = NULL)
}
\arguments{
\item{abap_data}{single-season ABAP data downloaded using \code{\link{getAbapData}}.}

\item{pentads}{an sf object returned by \code{\link{getRegionPentads}}. Defaults to \code{NULL}.}
}
\value{
A list containing data necessary for model fitting in \code{spOccupancy}. List elements are \code{y} (detection/non-detection data), \code{det.covs} (survey-level covariates), and \code{coords} (x and y centroids of pentads if spatial data are supplied).
}
\description{
This function transforms a raw ABAP data frame (returned by \code{\link{getAbapData}}) into an list which can be used to fit single-species occupancy models using \code{\link[spOccupancy]{spPGOcc}}(spatial) or \code{\link[spOccupancy]{PGOcc}}(non-spatial). The \code{spOccupancy} package fits single-species, multi-species, and integrated non-spatial and spatial occupancy models using Markov Chain Monte Carlo (MCMC).
}
\details{
The \code{\link[spOccupancy]{spPGOcc}} function takes \code{coords} as an argument with X and Y coordinates included for each site. Within the context of ABAP these \code{coords} are the centroid of each sampling pentad. In order to provide these spatial data to this function, simply use \code{\link{getRegionPentads}} and provide the same inputs for \code{.region_type} and \code{.region} that are specified in corresponding \code{\link{getAbapData}} call.\cr

If \code{pentads} are specified then the output can be automatically used in \code{\link[spOccupancy]{spPGOcc}}. If no spatial data are provided the output can be used in \code{\link[spOccupancy]{PGOcc}}.\cr

In addition to reformatting the detection/non-detection ABAP data for use in \code{spOccupancy} occupancy models, this function also extracts two survey-level covariates and adds them to the output list: \code{ hours} and \code{jday}. The \code{hours} variable is the total number of hours spent atlassing which is recorded on the pentad card and \code{jday} is the Julian day corresponding to the first day of atlassing for that card.
}
\examples{
abap_data <- getAbapData(.spp_code = 212,
                         .region_type = "province",
                         .region = "Eastern Cape",
                         .years = 2012)

abap_pentads <- getRegionPentads(.region_type = "province",
                                 .region = "Eastern Cape")


## Return list for spatial occupancy model
abapToSpOcc_single(abap_data, abap_pentads)

## List for non-spatial occupancy model
abapToSpOcc_single(abap_data)
}
\seealso{
\code{\link[spOccupancy]{spPGOcc}}, \code{\link[spOccupancy]{PGOcc}}
}
\author{
Dominic Henry \href{mailto:dominic.henry@gmail.com}{dominic.henry@gmail.com} \cr
Pachi Cervantes
}
