% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abapToUnmarked_multi.R
\name{abapToUnmarked_multi}
\alias{abapToUnmarked_multi}
\title{ABAP to unmarked}
\usage{
abapToUnmarked_multi(abap_data)
}
\arguments{
\item{abap_data}{ABAP data dowloaded using \link{getAbapData}.}
}
\value{
An \link{unmarkedMultFrame}
}
\description{
Transforms a raw ABAP data into multiseason unmarked-ready data
}
\examples{
abap_multi <- getAbapData(.spp = 212,
                          .region_type = "province",
                          .region = "Eastern Cape",
                          .years = c(2009,2010,2011,2012))

um_df <- abapToUnmarked_multi(abap_multi)
}
