% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EEcollectionToMultiband.R
\name{EEcollectionToMultiband}
\alias{EEcollectionToMultiband}
\title{Convert a GEE collection to multiband image}
\usage{
EEcollectionToMultiband(
  collection,
  dates,
  band,
  group_type,
  groups,
  reducer = NULL,
  unmask = FALSE
)
}
\arguments{
\item{collection}{Either a character string with the name of the collection
we want to use or a GEE collection produced with \code{ee$ImageCollection()}.
See \href{https://developers.google.com/earth-engine/datasets/catalog}{GEE catalog}.}

\item{dates}{A character vector with two elements c(start, end). Format must be
"yyyy-mm-dd".}

\item{band}{Select a specific band from the collection. Only one band from
the collection can be processed at a time.}

\item{group_type}{The type of grouping we want to perform. This has to be a
temporal grouping, either: 'day', 'week', 'month', or 'year'.}

\item{groups}{An integer vector with values for the group. e.g.
\code{2008:2019} will group data into years from 2008 to 2019.}

\item{reducer}{The summarizing function applied to each group. If reducer is
NULL, then the collection is transformed into a multiband image without
applying any reduction (each image goes to a band).}

\item{unmask}{GEE masks missing values, which means they are not used for
computing means, counts, etc. Sometimes we might want to avoid this behaviour
and use 0 instead of NA. If so, set unmask to TRUE.}
}
\value{

}
\description{
Convert a GEE collection to multiband image
}
