% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionPentads.R
\name{getRegionPentads}
\alias{getRegionPentads}
\title{Get SABAP2 pentads for a certain region}
\usage{
getRegionPentads(.region_type, .region)
}
\arguments{
\item{.region_type}{The type of region we are interested on.
Three options: "country", "province", "project" and "pentad".}

\item{.region}{A character string corresponding to the specific region we are
interested in. It can be either a country in Southern Africa, a South African
province, a project ("SABAP1" or "SABAP2") or a pentad code.}
}
\value{
A simple feature object with those pentads that are contained in the
region of interest. A pentads is considered contained if half of its area
lies within the region.
}
\description{
Get SABAP2 pentads for a certain region
}
\examples{
# getRegionPentads(.region_type = "country", .region = "South Africa")
}
