% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRegionPentads.R
\name{getRegionPentads}
\alias{getRegionPentads}
\title{Get SABAP2 pentads for a certain region}
\usage{
getRegionPentads(.country, .province = NULL, .path = NULL)
}
\arguments{
\item{.country}{A character string with the name of the country the region of
interest is at. Capitalize first letters of each word.}

\item{.province}{A character string. If provided it subsets the region of
interest to a specific province. Capitalize first letters of each word.}

\item{.path}{A directory where administrative boundaries layer should be found
or stored at.}
}
\value{
A simple feature object with those pentads that are contained in the
region of interest. A pentads is considered contained if half of its area
lies within the region. Note that the function uses raster::getData function,
which will download the administrative boundaries data once per session.
Alternatively, you may provide a path where this info will be stored and
retrieved from permanently.
}
\description{
Get SABAP2 pentads for a certain region
}
\examples{
# getRegionPentads(country = "South Africa", province = "North West")
}
